+++ src/index.ts
@@ -1,4 +1,0 @@
-// https://everythingfonts.com/ttf-to-woff - converter
-// https://caniuse.com/?search=woff
-import './styles.css';
-export * from './wxtile-js';